requirements.txt
flask-cors
pytorch
transformers
